control_server/requirements.txt
django==1.9.2
pyzmq==15.2.0
aiohttp==0.21.0
djangorestframework==3.3.2